+++ src/execution_plan/execution_plan_build/build_call_subquery.h
@@ -1,0 +1,12 @@
+/*
+ * Copyright Redis Ltd. 2018 - present
+ * Licensed under your choice of the Redis Source Available License 2.0 (RSALv2) or
+ * the Server Side Public License v1 (SSPLv1).
+ */
+
+// construct the execution-plan corresponding to a call {} clause
+void buildCallSubqueryPlan
+(
+	ExecutionPlan *plan,            // execution plan to add plan to
+	const cypher_astnode_t *clause  // call subquery clause
+);
