+++ src/execution_plan/ops/op_call_subquery.h
@@ -36,14 +36,14 @@ typedef struct Connector{
 typedef struct {
     OpBase op;
 
-    bool first;            // is this the first call to consume
-    bool is_eager;         // is the op eager
-    bool is_returning;     // is the subquery returning or not
-    OpBase *body;          // first op in the embedded execution-plan
-    OpBase *lhs;           // op from which records are pulled
-    Record r;              // current record consumed from lhs
-    Record *records;       // records aggregated by the operation
-    Connector connectors;  // connectors to the body (Args/ArgLists)
+    bool first;         // is this the first call to consume
+    bool is_eager;      // is the op eager
+    bool is_returning;  // is the subquery returning or not
+    OpBase *body;       // first op in the embedded execution-plan
+    OpBase *lhs;        // op from which records are pulled
+    Record r;           // current record consumed from lhs
+    Record *records;    // records aggregated by the operation
+    Connector feeders;  // connectors to the body (Args/ArgLists)
 } OpCallSubquery;
 
 // creates a new CallSubquery operation
